🚀 chore(package.json): update dependencies and scripts This commit updates the dependencies in the package.json file to their latest versions. It also updates the start:dev script to set the NODE_ENV environment variable to development. This is useful for development purposes as it allows the application to run in development mode.
🚀 chore(app.controller.ts): update logger injection
This commit updates the logger injection in the AppController to use the nest-winston package instead of the deprecated nestjs-pino package. It also updates the import statements to reflect the changes.

🚀 chore(app.module.ts): update imports and configurations
This commit updates the imports and configurations in the app.module.ts file. It updates the grpcOption import to reflect the new location of the file. It also updates the OpenTelemetryModule import to use the new version of the package. The LoggerModule import is removed as it is no longer needed. The WinstonModule is added to the imports and is

🔧 chore(auth-api): reorganize configuration files and update dependencies
This commit includes several changes to the configuration files of the auth-api service. The grpcOption file has been moved to the config folder and renamed to grpc.option.ts. The tracing configuration has been moved to its own file in the config folder. The winston configuration has also been moved to its own file in the config folder. Additionally, the dependencies for the project have been updated.

🔨 refactor(task.ts): make all properties optional
🔨 refactor(user.ts): make all properties optional
🔨 chore(tsconfig.json): update target to ES2022
The changes made to the task and user interfaces make all properties optional, which allows for more flexibility when creating and updating tasks and users. The target in the tsconfig.json file was updated to ES2022 to take advantage of the latest features in ECMAScript.